Module Math.Random

val random_seed : float -> unit

Set the random seed

val random : float -> float -> float

random a b picks a random number between a and b

val random_choices : float list -> float

random lst will choose a random element of lst

val random_gaussian : float -> float -> float

random_gaussian mean sd returns a random number generated from a Gaussian with mean mean and standard deviation sd