Module Shape.Curves

val bezier : float list -> unit

bezier points draws a bezier curve. List of parameters can be seen here

val bezier_detail : int -> unit

Sets the resolution of the Bezier curve -- default is 20

val bezier_point : float -> float -> float -> float -> float -> float

bezier_point a b c d t Given x or y co-ordinate values of control and anchor points of a bezier, it calculates the x or y coordinate of the bezier at position t

val bezier_tangent : float -> float -> float -> float -> float -> float

Evaluates the tangent to the Bezier at position t for points a, b, c, d -- Ref

val curve : float list -> unit

Draws a curve between two points -- Ref

val curve_detail : int -> unit

Sets the resolution of the curve -- default is 20 and min of 3

val curve_tightness : int -> unit

Sets how tight the curve fits to the vertex points -- default is 0.0

val curve_tangent : float -> float -> float -> float -> float -> float

Evaluates the curve tangent at position t for points a, b, c, d-- Ref

val curve_point : float -> float -> float -> float -> float -> float

Evaluates the curve at position t for points a, b, c, d-- Ref