Module Primitives.ThreeD

val plane : ?⁠detail_x:int -> ?⁠detail_y:int -> float -> float -> unit

plane w h will draw a plane with width w and height h. Optional detail_ for amount of segments defaulting to 16

val box : ?⁠detail_x:int -> ?⁠detail_y:int -> float -> float -> float -> unit

box w h d will draw a box with width w, height h and depth d. Optional detail_ for amount of segments defaulting to 16

val sphere : ?⁠detail_x:int -> ?⁠detail_y:int -> float -> unit

sphere r will draw a sphere with radius r. Optional detail_ for amount of segments defaulting to 16

val cylinder : ?⁠detail_x:int -> ?⁠detail_y:int -> ?⁠bot_cap:bool -> ?⁠top_cap:bool -> float -> float -> unit

cylinder r h will draw a cylinder with radius r and height h. Optional detail_ for amount of segments defaulting to 16. Optional _cap for disabling the caps on either end of the cylinder.

val cone : ?⁠detail_x:int -> ?⁠detail_y:int -> ?⁠cap:bool -> float -> float -> unit

cone r h will draw a cone with radius r and height h. Optional detail_ for amount of segments defaulting to 16. Optional cap to disable the capping of the cone.

val ellipsoid : ?⁠detail_x:int -> ?⁠detail_y:int -> float -> float -> float -> unit

ellipsoid rx ry rz will draw an ellipsoid with x-radius rx, y-radius ry and z-radius rz. Optional detail_ for amount of segments defaulting to 16.

val torus : ?⁠detail_x:int -> ?⁠detail_y:int -> float -> float -> unit

torus r tube will draw a torus with outer radius r and inner radius tube. Optional detail_ for amount of segments defaulting to 16.

val line : float -> float -> float -> float -> float -> float -> unit

line x1 y1 z1 x2 y2 z2 draws a line from (x1, y1, z1) to (x2, y2, z2)