Module P5__.Vector

type t = private Ojs.t

The type of P5 Vectors

val t_to_js : t -> Ojs.t

val t_of_js : Ojs.t -> t

val create_vector : float -> float -> ?⁠z:float -> unit -> t

val set : t -> float list -> unit

Set the elements of a vector -- Ref

val copy : t -> t

Copy a vector to a new vector

val get_x : t -> float

val get_y : t -> float

val get_z : t -> float

val add : t -> t -> t

Element-wise addition of two vectors

val sub : t -> t -> t

Element-wise subtraction of two vectors

val mult : t -> float -> t

mult t s scales a vector t by a scalar s

val rem : t -> t -> t

Element-wise remainder of two vectors

val div : t -> float -> t

div t s divides a vector t by a scalar s

val mag : t -> unit -> float

mag t () returns the magnitude of vector t

val mag_sq : t -> unit -> float

mag_sq t () returns the magnitude squared of vector t

val dot : t -> t -> t

dot a b calculates the dot-product of vectors a and b

val cross : t -> t -> t

cross a b calculates the cross-product of vectors a and b

val normalize : t -> t

Normalizes a vector -- gives it unit length

val limit : t -> float -> unit

Like normalize but for a supplied limit

val set_mag : t -> float -> unit

set_mag t len sets the magnitude of t to len

val heading : t -> unit -> float

Calculate the angle of rotation for a vector t

val rotate : t -> float -> unit

rotate t r rotates a vector t by r

val angle_between : t -> t -> float

angle_between a b calculates the angle between a and b

val lerp : t -> float -> float -> float -> float -> unit

lerp v x y z amt will linear interpolate v to (x, y, z) by amt where 0.0 <= amt <= 1.0

val reflect : t -> t -> unit

reflect v n reflects v about surface normal or line n

val array : t -> unit -> float array

Return a float array representation of the vector

val equals : t -> t -> bool

Equality check for vectors

module P5 : sig ... end