Module P5__Math.Calc

val abs : float -> float

abs f returns the absolute value of f

val ceil : float -> float

ceil f returns the ceiling value of f

val constain : float -> float -> float -> float

constrain n low high will keep n between low and high

val dist : float list -> float

Returns the distance between two points -- Ref

val exp : float -> float

exp n returns Euler's number raised to n

val floor : float -> float

floor n floors n

val lerp : float -> float -> float -> float

lerp a b amt linear interpolates between a and b by amt

val log : float -> float

log n calculates the natural logarithm of n

val mag : float -> float -> float

mag a b calculates the magnitudes of the vector (0, 0) to (a, b)

val map : float -> float -> float -> float -> float -> float

map v s1 f1 s2 f2 remaps a number v from one range s1 - f1 to a new range s2 - f2

val max : float list -> float

max [f1; f2...] finds the max element of the list

val min : float list -> float

min [f1; f2...] finds the min element of the list

val norm : float -> float -> float -> float

norm v s f normalises a number v from a range s to f

val pow : float -> float -> float

pow n e raises n to e

val round : float -> ?⁠deci:float -> unit -> float

round x () will round x, optionally specify to how many decimal places

val sq : float -> float

sq n squares n

val sqrt : float -> float

sqrt n takes the square root of n

val fract : float -> float

Calculates the fractional part of a number