Module P5__Shape.Vertex

val begin_contour : unit -> unit

Start a contour -- Ref

val end_contour : unit -> unit

End a contour -- Ref

val bezier_vertex : float list -> unit

Specifies vertex coordinates for Bezier curves -- Ref

val curve_vertex : ?⁠z:float -> float -> float -> unit

curve_vertex x y specifies vertex coordinates for curves -- Ref

val begin_shape : unit -> unit

begin_shape starts recording the vertices for more complex shapes -- Ref

val end_shape : unit -> unit

begin_shape stops recording the vertices for more complex shapes -- Ref

val vertex : float -> float -> z:float -> unit

vertex x y creates a new vertex at x and y, optionally specify the z coordinate

val stroke : float -> float -> float -> unit